fix: retry fulfillTransfer when hold account is busy
There was retry logic for submitting fulfillments in `src/models/transfers.js`,
but it was broken because the error code it  was looking for was being swapped
out for a different error by lower-level code in `src/lib/hold.js`.

This PR removes that retry logic, and instead moves it to `src/lib/db.js`, so
that it applies to all serializable transactions, not just to fulfillment submitting.

It also fixes the code that was masking the error code.